+++ components/User/ProductMenuList.tsx
@@ -1,7 +1,7 @@
-import { ChevronRightIcon } from '@chakra-ui/icons';
-import { Image, Flex, Text, Divider } from '@chakra-ui/react';
-import { useRouter } from 'next/router';
+import { Flex } from '@chakra-ui/react';
 import { Fragment } from 'react';
+
+import { ProductMenuItem } from '.';
 
 interface ProductMenu {
   iconUrl: string;
@@ -14,31 +14,14 @@ interface ProductMenuListProps {
 }
 
 const ProductMenuList = ({ productMenues }: ProductMenuListProps) => {
-  const router = useRouter();
-
   return (
     <Flex width="100%" direction="column" gap="12px" marginTop="21px">
-      {productMenues.map(({ iconUrl, title, routingUrl }, index) => (
+      {productMenues.map((productMenuProp, index) => (
         <Fragment key={index}>
-          <Flex
-            width="100%"
-            alignItems="center"
-            gap="17px"
-            onClick={() => router.push(routingUrl)}
-          >
-            <Image src={iconUrl} alt={`${title} 아이콘`} />
-            <Text
-              flexGrow="1"
-              fontFamily="Roboto"
-              fontStyle="normal"
-              fontWeight="400"
-              fontSize="16"
-            >
-              {title}
-            </Text>
-            <ChevronRightIcon justifySelf="flex-end" />
-          </Flex>
-          {index !== productMenues.length - 1 ? <Divider /> : undefined}
+          <ProductMenuItem
+            {...productMenuProp}
+            isLastItem={index === productMenues.length - 1}
+          />
         </Fragment>
       ))}
     </Flex>
