+++ components/User/index.tsx
@@ -1,3 +1,4 @@
 export { default as NoSellProducts } from './NoSellProducts';
 export { default as UserProfileInformation } from './UserProfileInformation';
 export { default as UserProfileEditButton } from './UserProfileEditButton';
+export { default as ProductMenuItem } from './ProductMenuItem';
